Make TableCreator class modular
